Move CubemapLayout out of decal code
